+++ src/main/java/fr/estecka/shiftingwares/ShiftingWaresMod.java
@@ -1,0 +1,38 @@
+package fr.estecka.shiftingwares;
+
+import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.gamerule.v1.CustomGameRuleCategory;
+import net.fabricmc.fabric.api.gamerule.v1.GameRuleFactory;
+import net.fabricmc.fabric.api.gamerule.v1.GameRuleRegistry;
+import net.minecraft.item.ItemStack;
+import net.minecraft.text.Text;
+import net.minecraft.util.Formatting;
+import net.minecraft.util.Identifier;
+import net.minecraft.village.TradeOffer;
+import net.minecraft.world.GameRules;
+import net.minecraft.world.GameRules.BooleanRule;
+import fr.estecka.shiftingwares.TradeLayouts.VanillaTradeLayout;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class ShiftingWaresMod
+implements ModInitializer
+{
+	static public final String MODID = "shifting-wares";
+	static public final Logger LOGGER = LoggerFactory.getLogger(MODID);
+
+	static public final CustomGameRuleCategory RULE_CATEGORY = new CustomGameRuleCategory(Identifier.of(MODID, "gamerules"), Text.translatable("gamerule.category.shiftingwares").formatted(Formatting.BOLD, Formatting.YELLOW));
+
+	static public final GameRules.Key<BooleanRule> WORKSTATION_RULE = GameRuleRegistry.register("shiftingWares.workstationProtection", RULE_CATEGORY, GameRuleFactory.createBooleanRule(true));
+	static public final GameRules.Key<BooleanRule> DAILY_RULE    = GameRuleRegistry.register("shiftingWares.dailyReroll",    RULE_CATEGORY, GameRuleFactory.createBooleanRule(true));
+	static public final GameRules.Key<BooleanRule> DEPLETED_RULE = GameRuleRegistry.register("shiftingWares.depleteReroll",  RULE_CATEGORY, GameRuleFactory.createBooleanRule(true));
+	static public final GameRules.Key<BooleanRule> MAP_RULE      = GameRuleRegistry.register("shiftingWares.allowMapReroll", RULE_CATEGORY, GameRuleFactory.createBooleanRule(false));
+
+	static public final TradeOffer PLACEHOLDER_TRADE = new TradeOffer(ItemStack.EMPTY, ItemStack.EMPTY, ItemStack.EMPTY, 0, 0, 0, 0, 0);
+
+	@Override
+	public void onInitialize() {
+		// Static initialization
+		LOGGER.info("Trade-Rebalance support is {} for this version of minecraft.", VanillaTradeLayout.IS_EXP_TRADE_AVAILABLE?"enabled":"disabled");
+	}
+}