+++ src/main/java/tk/estecka/shiftingwares/api/IHasItemCache.java
@@ -1,9 +1,0 @@
-package tk.estecka.shiftingwares.api;
-
-/**
- * Implemented by VillagerEntity
- */
-public interface IHasItemCache
-{
-	PersistentItemCache	shiftingwares$GetItemCache();
-}